backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.15"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "aiohttp (>=3.13.3,<4.0.0)",
    "async-lru (>=2.0.5,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
